Created `buildServer()` function
This improves separation of concerns and makes testing easier.
